refactor: decouple component analysis algorithms from AbsGraph
