+++ w9_pathfinding/src/core.cpp
@@ -54,126 +54,6 @@ bool AbsGraph::is_valid_path(Path& path) {
     return true;
 }
 
-vector<int> AbsGraph::find_component_(vector<bool> &visited, int start) {
-    visited[start] = true;
-    vector<int> component = {start};
-
-    vector<int> stack = {start};
-    while (!stack.empty()) {
-        int x = stack.back();
-        stack.pop_back();
-        for (auto& [n, cost] : get_neighbors(x)) {
-            if (!visited[n]) {
-                visited[n] = true;
-                component.push_back(n);
-                stack.push_back(n);
-            }
-        }
-    }
-
-    return component;
-}
-
-vector<vector<int>> AbsGraph::find_components() {
-    if (is_directed_graph())
-        throw std::runtime_error("find_components only works for an undirected graph");
-
-    int graph_size = size();
-    vector<bool> visited(graph_size, false);
-    vector<vector<int>> components;
-
-    int offset = 0;
-    while (true) {
-
-        int start = -1;
-        for (int i = offset; i < graph_size; i++) {
-            if (!visited[i]) {
-                start = i;
-                offset = i + 1;
-                break;
-            }
-        }
-
-        if (start == -1)
-            break;
-
-        components.push_back(find_component_(visited, start));
-    }
-
-    return components;
-}
-
-void dfs_with_order_(AbsGraph *graph, vector<bool> &visited, vector<int> &order, int start) {
-    visited[start] = true;
-    for (auto& [n, cost] : graph->get_neighbors(start, true)) {
-        if (!visited[n])
-            dfs_with_order_(graph, visited, order, n);
-    }
-    order.push_back(start);
-}
-
-vector<int> dfs_sort_(AbsGraph *graph) {
-    int graph_size = graph->size();
-
-    vector<bool> visited(graph_size, false);
-    vector<int> order;
-
-    int offset = 0;
-    while (true) {
-
-        int start = -1;
-        for (int i = offset; i < graph_size; i++) {
-            if (!visited[i]) {
-                start = i;
-                offset = i + 1;
-                break;
-            }
-        }
-
-        if (start == -1)
-            break;
-
-        dfs_with_order_(graph, visited, order, start);
-    }
-
-    return order;
-}
-
-vector<vector<int>> AbsGraph::find_scc() {
-    if (!is_directed_graph()) {
-        throw std::runtime_error("find_scc only works for a directed graph");
-    }
-
-    // Kosaraju's algorithm
-
-    vector<int> order = dfs_sort_(this);
-    std::reverse(order.begin(), order.end());
-
-    int graph_size = size();
-    vector<bool> visited(graph_size, false);
-    vector<vector<int>> scc;
-
-    int offset = 0;
-    while (true) {
-
-        int start = -1;
-        for (int i = offset; i < graph_size; i++) {
-            if (!visited[order[i]]) {
-                start = order[i];
-                offset = i + 1;
-                break;
-            }
-        }
-
-        if (start == -1)
-            break;
-
-        scc.push_back(find_component_(visited, start));
-    }
-
-    return scc;
-}
-
 bool AbsGraph::adjacent(int v1, int v2) {
     for (auto &[node_id, cost] : get_neighbors(v1, false, true)) {
         if (node_id == v2) {
@@ -261,17 +141,6 @@ vector<double> AbsGrid::get_pause_weights() const {
 
     vector<double> weights(size(), pause_weight_);
     return weights;
-}
-
-vector<vector<int>> AbsGrid::find_components() {
-    vector<vector<int>> components = AbsGraph::find_components();
-    vector<vector<int>> components_without_walls;
-    for (vector<int> &x : components) {
-        if (x.size() == 1 && has_obstacle(x[0]))
-            continue;
-        components_without_walls.push_back(x);
-    }
-    return components_without_walls;
 }
 
 void ensure_path_length(Path& path, int length) {
