+++ w9_pathfinding/src/include/components.h
@@ -1,0 +1,8 @@
+#include "core.h"
+
+// Finds connected components in an undirected graph
+vector<vector<int>> find_components(AbsGraph* graph);
+
+
+// Finds Strongly Connected Components (SCC) in a directed graph
+vector<vector<int>> find_scc(AbsGraph* graph);
